Rework game input handling and improve code flow
There's a lot going on here since everything is intertwined so i'll
explain this step by step.

1. Changed lane variable from a string to an enum value.
Enum values are named constants, backed by an integer.  They're used in
situations where you have a set number of unique values where the actual
content of the value may not matter.  In this case, our Lane enum has 3
values, Top, Mid, and Bot.  This is preferable to using strings for
arbitrary values, as mentioned in the comment.  The type for the lane
variable was also changed in each unit file.

2. Reworked input handling
Like with the main menu change, it's better to use the input events
rather than polling for key presses in this situation.  Matching based
on keycode constants is preferable to polling with if chains.  The code
to set the selected unit and spawn units with mana checks+handling have
also been moved to their own functions to reduce redundancy.  The mouse
press event is in a different function since the other one is for keys.

3. Changed how mana is restored
This is explained in detail in the comment, but in short, it uses the
physics delta to add mana at the correct rate.

4. Moved ai code to its own function
A simple organizational change.  Removes redundant checking as well.

5. Misc code formatting stuff
Adding types and line breaks between logical code sections helps with
readability.