More generic unit progress and refactor weapons into equipment
I had a bit of a design bump doing ranged weapons and needed to change
the system a bit.  Now instead of ranged weapons trying to create and
shoot themselves, there's ranged weapons and projectiles the weapons
shoot.  This makes much more sense and I feel silly for having to have
had to figure it out the hard way.  Also did more work on units.